Use cumulative average for BPM detection
The BPM detection was using just the delta between two taps. This isn't
particularly accurate as people tend to have some variation when tapping
rhythms. This code change uses a cumlative(-ish) average to calculate
the BPM.

Unlike a true cumulative average, where the average is calculated as
`bpm_average = (delta + (current_bpm_average * (k-1)))/k` where `delta`
is the time between the two most recent taps and `k` is the number of
data points we have seen so far, this code limits `k` to `3`. This
allows the BPM average to easily recover when users are a bit off when
tapping and prevents outliers from having too much influence over the
average.

The limit of `3` was chosen because manual testing showed that higher
values made converging to a BPM take too long when outliers existed
(e.g. a user rushes or drags a few taps), and lower values were
ineffective in accounting for human error.

Another approach would be to use a rolling average, but that would
require slightly more memory and may not be as effective in combatting
human error